% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arclabels.R
\name{arclabels}
\alias{arclabels}
\alias{arclabels.default}
\title{Add arc labels to tips of a phylogeny; works for non-monophyletic groups and single tip lineages.}
\usage{
arclabels(tips, ...)

\method{arclabels}{default}(phy = NULL, tips, text,
  plot_singletons = TRUE, ln.offset = 1.02, lab.offset = 1.06,
  cex = 1, orientation = "horizontal", ...)
}
\arguments{
\item{tips}{A character vector (or a list ?) with the names of the tips that belong to the clade or group. If multiple groups are going to be plotted, tips must be given in the form of a list.}

\item{...}{optional arguments for \code{arc.cladelabels}.}

\item{phy}{An object of class phylo.}

\item{text}{A character vector indicating the desired text to label the arcs.}

\item{plot_singletons}{Boolean. If TRUE (default), it will add arcs (and labels) to single tip lineages too. If FALSE, no arc or labels will be plotted over that tip..}

\item{ln.offset}{line offset (as a function of total tree height) for \code{arc.cladelabels}.}

\item{lab.offset}{label offset for \code{arc.cladelabels}.}

\item{cex}{character expansion factor.}

\item{orientation}{orientation of the text. Can be \code{orientation = "vertical"} (the default) or \code{"horizontal"}.}
}
\value{
\code{NULL}
}
\description{
Add arc labels to tips of a phylogeny; works for non-monophyletic groups and single tip lineages.
}
