% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparetrees.R
\name{CompareTrees}
\alias{CompareTrees}
\title{Compare two phylogenetic trees}
\usage{
CompareTrees(trees)
}
\arguments{
\item{trees}{A character vector of two newick trees}
}
\value{
A named list with are_same_tree property set to either TRUE or FALSE
}
\description{
Compare two phylogenetic trees
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md}
}
