% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel.R
\name{data_get_tree}
\alias{data_get_tree}
\title{Function to get tree for a data set.}
\usage{
data_get_tree(data, tnrs_source = "otol", tree_source = "otol",
  prune = TRUE, summary_format = "phylo_biggest", ...)
}
\arguments{
\item{data}{Data.frame with species names as rownames}

\item{tnrs_source}{Source for taxonomic name resolution; options are "otol" and "gnr". If set to NULL, assumes names are fine as is}

\item{tree_source}{Source for tree; options are "otol", "phylomatic", "datelife".}

\item{prune}{If TRUE, delete taxa to matching sets only}

\item{summary_format}{What format to return from datelife}

\item{...}{Other options to pass to datelife::datelife_search}
}
\value{
list with a phy and a data object, both pruned to the same taxon set
}
\description{
A common use case is having a set of data for a set of species and wanting to
get a tree for those species. This involves resolving to the same taxonomy,
getting a tree for those species, and then (optionally) pruning the data and
tree to the same set of taxa (this is optional: there are approaches to
make up trait data or phylogenetic placement in the absence of information).
}
\details{
For sources of trees, besides the standard ones (Open Tree of Life, Phylomatic),
datelife is also an option. For this, you will have had to install the
datelife package (which is only suggested, not required, by this code).
It will return the chronogram in the tree store with the most overlap to
your set of taxa by default, but you can change options with summary_format;
see ?datelife::datelife_search for more info
}
