% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{file_get_scientific_names}
\alias{file_get_scientific_names}
\title{Function to pull scientific names from file}
\usage{
file_get_scientific_names(file_name, search_engine = 0)
}
\arguments{
\item{file_name}{The file path and name to extract names from}

\item{search_engine}{1 to use TaxonFinder, 2 to use NetiNeti, 0 to use both}
}
\value{
A vector of scientific names
}
\description{
This uploads a file (a PDF, Microsoft Word document, plain text file, etc.) and extracts all scientific names from it. For example, you can input a PDF of a scientific article and it will return all the scientific names in that article.
}
\details{
It requires that curl is installed on your system.
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription}
}
