% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ready_lists.R
\docType{data}
\name{flower_plant_fams}
\alias{flower_plant_fams}
\title{Flowering plants families from Open Tree Taxonomy}
\format{A character vector}
\source{
\url{https://tree.opentreeoflife.org/about/taxonomy-version/ott3.0}
}
\usage{
flower_plant_fams
}
\description{
Flowering plants families from Open Tree Taxonomy
}
\details{
How was the data were obtained or generated
flower_plant_fams <- datelife::get_ott_children(ott_ids = 99252, ott_rank = "family")
flower_plant_fams <- flower_plant_fams[[1]]
flower_plant_fams <- rownames(flower_plant_fams)[as.character(flower_plant_fams[,"rank"]) == "family"]
usethis::use_data(flower_plant_fams)
}
\author{
Luna L. Sanchez-Reyes \email{lsanche7@utk.edu}
Brian O'Meara \email{bomeara@utk.edu}
}
\keyword{magnoliophyta}
\keyword{otol}
