% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_base_url}
\alias{get_base_url}
\title{Get Phylotastic base URL
return The URL for the phylotastic server}
\usage{
get_base_url()
}
\description{
Get Phylotastic base URL
return The URL for the phylotastic server
}
