% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{GetBaseURL}
\alias{GetBaseURL}
\title{Get Phylotastic base URL
return The URL for the phylotastic server}
\usage{
GetBaseURL()
}
\description{
Get Phylotastic base URL
return The URL for the phylotastic server
}
