% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspecies.R
\name{GetCountrySpeciesFromTaxon}
\alias{GetCountrySpeciesFromTaxon}
\title{Get all species filtered by country from a taxon}
\usage{
GetCountrySpeciesFromTaxon(taxon, country, filters = c("environmental",
  "sp\\\\.", "cf\\\\."))
}
\arguments{
\item{taxon}{A taxon to get subset of species that are established in a particular country}

\item{country}{A country name where species of the input taxon are established.}

\item{filters}{A character vector of strings to exclude}
}
\value{
A vector of names
}
\description{
Get all species filtered by country from a taxon
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the rotl package, another interface to Open Tree of Life
}
