% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspeciesinfo.R
\name{GetImageDataSpecies}
\alias{GetImageDataSpecies}
\title{Get image metadata of a list of species}
\usage{
GetImageDataSpecies(species)
}
\arguments{
\item{species}{A vector of names}
}
\value{
A data frame of image metadata(image urls, license info etc.) of species
}
\description{
Get image metadata of a list of species
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or \url{https://eol.org/api/}
}
