% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspeciesinfo.R
\name{GetInfoSpecies}
\alias{GetInfoSpecies}
\title{Get information of a list of species}
\usage{
GetInfoSpecies(species)
}
\arguments{
\item{species}{A vector of names}
}
\value{
A data frame of info about species
}
\description{
Get information of a list of species
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or \url{https://eol.org/api/}
}
