% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listspecies.R
\name{GetListSpecies}
\alias{GetListSpecies}
\title{Get existing list/lists of species}
\usage{
GetListSpecies(userid, accesstoken, listid, verbose = FALSE, content = TRUE)
}
\arguments{
\item{userid}{A valid gmail address of the user}

\item{accesstoken}{Access token of the gmail address}

\item{listid}{An integer id of the list to retrieve}

\item{verbose(optional)}{By default FALSE and shows minimal meta-data of the list.}

\item{content(optional)}{By default TRUE and shows the species collection of the list}
}
\value{
An existing list with metadata and content based on parameters
}
\description{
Get existing list/lists of species
}
\examples{
  userid = "abusalehmdtayeen@gmail.com"
  accesstoken = "ya29..zQLmLjbyujJjwV6RVSM2sy-mkeaKu-9"
  listid = 12
  verbose = True
  content = False
  GetListSpecies(userid, accesstoken, listid, verbose, content)
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md}
}
