% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtrees.R
\name{GetOToLTree}
\alias{GetOToLTree}
\title{Get OToL induced subtree}
\usage{
GetOToLTree(taxa)
}
\arguments{
\item{taxa}{The vector of names, already resolved to match OToL taxa}
}
\value{
A phylo object
}
\description{
Get OToL induced subtree
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the rotl package, another interface to Open Tree of Life
}
