% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtrees.R
\name{GetPhylomaticTree}
\alias{GetPhylomaticTree}
\title{Get phylomatic subtree}
\usage{
GetPhylomaticTree(taxa)
}
\arguments{
\item{taxa}{The vector of names, already resolved}
}
\value{
A newick string
}
\description{
Get phylomatic subtree
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the interface of phylomatic \url{http://phylodiversity.net/phylomatic/}
}
