% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{GetScientificNamesFromText}
\alias{GetScientificNamesFromText}
\title{Function to pull scientific names from text}
\usage{
GetScientificNamesFromText(text, search.engine = 0)
}
\arguments{
\item{text}{The text string to extract names from}

\item{search.engine}{1 to use TaxonFinder, 2 to use NetiNeti, 0 to use both}
}
\value{
A vector of scientific names
}
\description{
Function to pull scientific names from text
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md}
}
