% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtrees.R
\name{GetTaxonomicTree}
\alias{GetTaxonomicTree}
\title{Get taxonomic tree from the NCBI}
\usage{
GetTaxonomicTree(taxa)
}
\arguments{
\item{taxa}{The vector of names, already resolved}
}
\value{
A newick string
}
\description{
Get taxonomic tree from the NCBI
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the interface of phyloT \url{http://phylot.biobyte.de/}
}
