% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparetrees.R
\name{newick_compare}
\alias{newick_compare}
\title{Compare two phylogenetic trees}
\usage{
newick_compare(trees)
}
\arguments{
\item{trees}{A multiphylo object of two newick trees}
}
\value{
A named list with are_same_tree property set to either TRUE or FALSE
}
\description{
Compare two phylogenetic trees
}
\examples{
are_same <- newick_compare(c(ape::rcoal(5), ape::rcoal(5)))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription}
}
