% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparetrees.R
\name{phylo_compare}
\alias{phylo_compare}
\title{Compare two phylogenetic trees}
\usage{
phylo_compare(phylo1, phylo2 = NULL)
}
\arguments{
\item{phylo1}{A multiphylo object with two trees to be compared between them or a single phylo object to be compared to phylo2}

\item{phylo2}{A phylo object to be compared to phylo1}
}
\value{
A named list with are_same_tree property set to either TRUE or FALSE
}
\description{
Compare two phylogenetic trees
}
\examples{
are_same <- phylo_compare(c(ape::rcoal(5), ape::rcoal(5)))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription}
}
