% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listspecies.R
\name{remove_species_from_list}
\alias{remove_species_from_list}
\title{Remove an existing list of species}
\usage{
remove_species_from_list(userid, access_token, list_id)
}
\arguments{
\item{userid}{A valid gmail address of the user}

\item{access_token}{Access token of the gmail address}

\item{list_id}{An integer id of the list to retrieve}
}
\value{
A list with the id of the list removed
}
\description{
Remove an existing list of species
}
\examples{
  # This gives you the syntax, but since the access token expires after one hour,
  # this particular example will not work.
  \dontrun{
  userid = "abusalehmdtayeen@gmail.com"
  access_token = "ya29..zQLmLjbyujJjwV6RVSM2sy-mkeaKu-9"
  list_id = 12
  remove_species_from_list(userid, access_token, list_id)
  }
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription}
}
