% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listspecies.R
\name{RemoveListSpecies}
\alias{RemoveListSpecies}
\title{Remove an existing list of species}
\usage{
RemoveListSpecies(userid, accesstoken, listid)
}
\arguments{
\item{userid}{A valid gmail address of the user}

\item{accesstoken}{Access token of the gmail address}

\item{listid}{An integer id of the list to retrieve}
}
\value{
A list with the id of the list removed
}
\description{
Remove an existing list of species
}
\examples{
  userid = "abusalehmdtayeen@gmail.com"
  accesstoken = "ya29..zQLmLjbyujJjwV6RVSM2sy-mkeaKu-9"
  listid = 12
  RemoveListSpecies(userid, accesstoken, listid)
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md}
}
