% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listspecies.R
\name{ReplaceListSpecies}
\alias{ReplaceListSpecies}
\title{Replace a list of species}
\usage{
ReplaceListSpecies(userid, accesstoken, listid, speciesobj)
}
\arguments{
\item{userid}{A valid gmail address of the user}

\item{accesstoken}{Access token of the gmail address}

\item{listid}{An integer id of the list to be modified}

\item{speciesObj}{A species object to replace with}
}
\value{
A list with the old species and new species list
}
\description{
Replace a list of species
}
\examples{
  userid = "abusalehmdtayeen@gmail.com"
  accesstoken = "ya29..zQLmLjbyujJjwV6RVSM2sy-mkeaKu-9"
  listid = 12
  speciesObj = list( list(family="",scientific_name="Aix sponsa",scientific_name_authorship="", vernacular_name="Wood Duck",phylum="",nomenclature_code="ICZN",order="Anseriformes",class="") )
  ReplaceListSpecies(userid, accesstoken, listid, speciesObj)
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md}
}
