% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{ResolveNamesWithOToL}
\alias{ResolveNamesWithOToL}
\title{Resolve Scientific Names with Open Tree TNRS}
\usage{
ResolveNamesWithOToL(taxa)
}
\arguments{
\item{taxa}{The vector of names}
}
\value{
A vector of corrected names. THE ORDER MAY NOT CORRESPOND TO YOUR INPUT ORDER.
}
\description{
Resolve Scientific Names with Open Tree TNRS
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the rotl package, another interface to Open Tree of Life, or the taxize package for name resolution in general.
}
