% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphylotastic.R
\docType{package}
\name{rphylotastic-package}
\alias{rphylotastic}
\alias{rphylotastic-package}
\title{rphylotastic: An R interface to 'Phylotastic' Web Services}
\description{
This wraps the 'Phylotastic' services APIs described on Web Services at <www.phylotastic.org>. The main use case is to return a phylogenetic tree for a set of species, but the services also include ways to extract species names from web pages, perform taxonomic name resolution, retrieve a list of all descendant species of a taxon, find images of a species, and more.
}
\details{
See `vignette("rphylotastic", package = "rphylotastic")` for an overview
of the package and its utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phylotastic/rphylotastic}
  \item \url{www.phylotastic.org}
  \item Report bugs at \url{https://github.com/phylotastic/rphylotastic/issues}
}

}
\author{
\strong{Maintainer}: Brian O'Meara \email{omeara.brian@gmail.com}

Authors:
\itemize{
  \item Abu Saleh Md Tayeen \email{abusalehmdtayeen@gmail.com}
  \item Luna L. Sanchez Reyes \email{sanchez.reyes.luna@gmail.com}
}

}
