% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separatedarktaxa.R
\name{SeparateDarkTaxa}
\alias{SeparateDarkTaxa}
\title{Separate dark from known taxa}
\usage{
SeparateDarkTaxa(taxon, filters = c("environmental", "sp\\\\.", "cf\\\\."))
}
\arguments{
\item{taxon}{A taxon to get all species for}

\item{filters}{A character vector of strings to exclude}
}
\value{
A vector of names
}
\description{
Separate dark from known taxa
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the rotl package, another interface to Open Tree of Life
}
