% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separatedarktaxa.R
\name{SeparateDarkTaxaDB}
\alias{SeparateDarkTaxaDB}
\title{Separate dark from known taxa on another database}
\usage{
SeparateDarkTaxaDB(taxon, filters = c("environmental", "sp\\\\.", "cf\\\\."),
  db = "col")
}
\arguments{
\item{taxon}{A taxon to get all species for}

\item{filters}{A character vector of strings to exclude}

\item{db}{See ?taxize::downstream for options}
}
\value{
A vector of names
}
\description{
Separate dark from known taxa on another database
}
