% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separatedarktaxa.R
\name{SeparateDarkTaxaGenbank}
\alias{SeparateDarkTaxaGenbank}
\title{Separate dark from known taxa on another database}
\usage{
SeparateDarkTaxaGenbank(taxon, filters = c("environmental", "sp\\\\.",
  "cf\\\\."))
}
\arguments{
\item{taxon}{A taxon to get all species for}

\item{filters}{A character vector of strings to exclude}
}
\value{
A vector of names
}
\description{
Separate dark from known taxa on another database
}
