% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separatedarktaxa.R
\name{SeparateDarkTaxaOToL}
\alias{SeparateDarkTaxaOToL}
\title{Separate dark from known taxa on OpenTree of Life}
\usage{
SeparateDarkTaxaOToL(taxon, filters = c("environmental", "sp\\\\.", "cf\\\\.",
  "uncultured", "unidentified", " clone", " enrichment"))
}
\arguments{
\item{taxon}{A taxon to get all species for}

\item{filters}{A character vector of strings to exclude}
}
\value{
A list containing a vector of dark names, a vector of known names, and fraction.dark
}
\description{
Separate dark from known taxa on OpenTree of Life
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the rotl package, another interface to Open Tree of Life
}
