% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspeciesinfo.R
\name{species_get_image_data}
\alias{species_get_image_data}
\title{Get image metadata of a list of species}
\usage{
species_get_image_data(species)
}
\arguments{
\item{species}{A vector of names}
}
\value{
A data frame of image metadata(image urls, license info etc.) of species
}
\description{
Get image metadata of a list of species
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or \url{https://eol.org/api/}
}
