% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolvescientificnames.R
\name{taxa_common_to_scientific}
\alias{taxa_common_to_scientific}
\title{Convert common names to scientific names.}
\usage{
taxa_common_to_scientific(taxa, service = "NCBI", multiple = FALSE)
}
\arguments{
\item{taxa}{A character vector of common names.}

\item{service}{Which service to use: NCBI, ITIS, or TROPICOS}

\item{multiple}{If TRUE, then the service will return multiple matches (if available) for each common name in the input list.}
}
\value{
A vector of scientific names. Output order may not correspond to input order.
}
\description{
Convert common names to scientific names.
}
\examples{
taxa <- c("blue whale", "swordfish", "killer whale")
common <- taxa_common_to_scientific(taxa)
print(common)
}
\seealso{
taxize package for name resolution in general and its sci2comm function.
}
