% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolvescientificnames.R
\name{taxa_convert_common_to_scientific}
\alias{taxa_convert_common_to_scientific}
\title{Convert common names to scientific names}
\usage{
taxa_convert_common_to_scientific(taxa, service = "NCBI",
  multiple = FALSE)
}
\arguments{
\item{taxa}{The vector of names}

\item{service}{Which service to use: NCBI, ITIS, or TROPICOS}

\item{multiple}{If TRUE, then the service will return multiple matches (if available) for each common name in the input list.}
}
\value{
A vector of correct names. THE ORDER MAY NOT CORRESPOND TO YOUR INPUT ORDER.
}
\description{
Convert common names to scientific names
}
\examples{
taxa <- c("blue whale", "swordfish","killer whale")
common <- taxa_convert_common_to_scientific(taxa)
print(common)
}
\seealso{
taxize package for name resolution in general and its sci2comm function.
}
