% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abovespecies.R
\name{taxa_get_above_species}
\alias{taxa_get_above_species}
\title{Get above species taxon names only from a vector of taxon names.}
\usage{
taxa_get_above_species(taxa)
}
\arguments{
\item{taxa}{A character vector of taxon names.}
}
\value{
A character vector of taxon names above species level.
}
\description{
Get above species taxon names only from a vector of taxon names.
}
\examples{
taxa_get_above_species("Vulpes_vulpes"))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the rotl package, another interface to Open Tree of Life
}
