% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtrees.R
\name{taxa_get_phylomatic_tree}
\alias{taxa_get_phylomatic_tree}
\title{Get phylomatic subtree}
\usage{
taxa_get_phylomatic_tree(taxa)
}
\arguments{
\item{taxa}{The vector of names, already resolved}
}
\value{
A newick string
}
\description{
Get phylomatic subtree
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the interface of phylomatic \url{http://phylodiversity.net/phylomatic/}
}
