% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtrees.R
\name{taxa_get_taxonomic_tree}
\alias{taxa_get_taxonomic_tree}
\title{Get taxonomic tree from the NCBI}
\usage{
taxa_get_taxonomic_tree(taxa)
}
\arguments{
\item{taxa}{The vector of names, already resolved}
}
\value{
A phylo object
}
\description{
Get taxonomic tree from the NCBI
}
\examples{
taxa <- c("Setophaga striata", "Setophaga magnolia",
     "Setophaga angelae", "Setophaga plumbea",
     "Setophaga virens")
phy <- taxa_get_taxonomic_tree(taxa)
plot(phy)
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the interface of phyloT \url{http://phylot.biobyte.de/}
}
