% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducespecies.R
\name{taxa_reduce_species}
\alias{taxa_reduce_species}
\title{Reduce species from a vector of taxon names. It will keep the most inclusive lineage above the species level.}
\usage{
taxa_reduce_species(taxa)
}
\arguments{
\item{taxa}{A character vector of taxon names.}
}
\value{
A character vector of lineage names above the species level.
}
\description{
Reduce species from a vector of taxon names. It will keep the most inclusive lineage above the species level.
}
\examples{
taxon_reduce_species("Vulpes_vulpes"))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the rotl package, another interface to Open Tree of Life
}
