% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolvescientificnames.R
\name{taxa_resolve_names_with_gnr}
\alias{taxa_resolve_names_with_gnr}
\title{Resolve Scientific Names with GNR TNRS}
\usage{
taxa_resolve_names_with_gnr(taxa)
}
\arguments{
\item{taxa}{The vector of names}
}
\value{
A vector of correct names. THE ORDER MAY NOT CORRESPOND TO YOUR INPUT ORDER.
}
\description{
Resolve Scientific Names with GNR TNRS
}
\details{
Mispelled or incorrect names will be dropped.
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the rotl package, another interface to Open Tree of Life, or the taxize package for name resolution in general.
}
