% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspecies.R
\name{taxon_get_species}
\alias{taxon_get_species}
\title{Get all species from a taxon from Open Tree of Life taxonomy.}
\usage{
taxon_get_species(taxon, filters = c("environmental", "sp\\\\.",
  "cf\\\\."))
}
\arguments{
\item{taxon}{A character vector of length 1. Specify the taxon name to get all species for. If vector is longer than 1, it will only take the first element and ignore all other names.}

\item{filters}{A character vector of strings to exclude.}
}
\value{
A character vector of species names.
}
\description{
Get all species from a taxon from Open Tree of Life taxonomy.
}
\examples{
print(taxon_get_species("Vulpes"))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the rotl package, another interface to Open Tree of Life
}
