% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspecies.R
\name{taxon_get_species_from_country}
\alias{taxon_get_species_from_country}
\title{Get all species filtered by country from a taxon}
\usage{
taxon_get_species_from_country(
  taxon,
  country,
  filters = c("environmental", "sp\\\\.", "cf\\\\.")
)
}
\arguments{
\item{taxon}{A character vector of length 1. Specify the taxon name to get a subset of species that are established in a particular country}

\item{country}{A country name where species of the input taxon are established.}

\item{filters}{A character vector of strings to exclude}
}
\value{
A vector of names
}
\description{
Get all species filtered by country from a taxon
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the rotl package, another interface to Open Tree of Life
}
