% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspecies.R
\name{taxon_get_species_with_genome}
\alias{taxon_get_species_with_genome}
\title{Get all species that have genome sequence in NCBI from a taxon}
\usage{
taxon_get_species_with_genome(taxon, filters = c("environmental", "sp\\\\.",
  "cf\\\\."))
}
\arguments{
\item{taxon}{A taxon to get subset of species having genome sequence}

\item{filters}{A character vector of strings to exclude}
}
\value{
A vector of names
}
\description{
Get all species that have genome sequence in NCBI from a taxon
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the rotl package, another interface to Open Tree of Life
}
