% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listspecies.R
\name{update_species_in_list}
\alias{update_species_in_list}
\title{Update metadata of a list of species}
\usage{
update_species_in_list(userid, accesstoken, listid, listObj)
}
\arguments{
\item{userid}{A valid gmail address of the user}

\item{accesstoken}{Access token of the gmail address}

\item{listid}{An integer id of the list to be modified}

\item{listObj}{A list object to update with}
}
\value{
A list with modified list metadata
}
\description{
Update metadata of a list of species
}
\examples{
  userid = "abusalehmdtayeen@gmail.com"
  accesstoken = "ya29..zQLmLjbyujJjwV6RVSM2sy-mkeaKu-9"
  listid = 12
  listObj = list(list_description="A sublist on the bird species",
            list_keywords=c("bird","Everglades"))
  update_species_in_list(userid, accesstoken, listid, listObj)
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md}
}
