% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{url_get_scientific_names}
\alias{url_get_scientific_names}
\title{Function to pull scientific names from web pages}
\usage{
url_get_scientific_names(URL, search.engine = 0)
}
\arguments{
\item{URL}{The URL to extract names from}

\item{search.engine}{1 to use TaxonFinder, 2 to use NetiNeti, 0 to use both}
}
\value{
A vector of scientific names
}
\description{
Function to pull scientific names from web pages
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md}
}
