% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{url_get_scientific_names_from_GNRD2}
\alias{url_get_scientific_names_from_GNRD2}
\title{Function to pull scientific names from web pages}
\usage{
url_get_scientific_names_from_GNRD2(
  URL,
  search_engine = 0,
  above_species = FALSE
)
}
\arguments{
\item{URL}{The URL to extract names from. Can be a pdf url.}

\item{search_engine}{1 to use TaxonFinder, 2 to use NetiNeti, 0 to use both}

\item{above_species}{Boolean. Default to FALSE. If TRUE it will only return scientific names above the species level.}
}
\value{
A vector of scientific names. It returns unique matches.
}
\description{
Function to pull scientific names from web pages
}
\examples{
# get scientific names from a wikipedia web page:
url_get_scientific_names(URL = "https://en.wikipedia.org/wiki/Plain_pigeon")
# get scientific names from a pdf URL:
url_get_scientific_names(URL =
 "http://darwin-online.org.uk/converted/pdf/1897_Insectivorous_F1229.pdf")
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription}
}
