% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspecies.R
\name{GetSpeciesFromTaxon}
\alias{GetSpeciesFromTaxon}
\title{Get all species from a taxon}
\usage{
GetSpeciesFromTaxon(taxon, filters = c("environmental", "sp."))
}
\arguments{
\item{taxon}{A taxon to get all species for}

\item{filters}{A character vector of strings to exclude}
}
\value{
A vector of names
}
\description{
Get all species from a taxon
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/blob/master/ServiceDescription/PhyloServicesDescription.md} or the rotl package, another interface to Open Tree of Life
}
