% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separatedarktaxa.R
\name{SeparateDarkTaxaGenbank}
\alias{SeparateDarkTaxaGenbank}
\title{Separate dark from known taxa on another database}
\usage{
SeparateDarkTaxaGenbank(taxon, filters = c("environmental", "sp\\\\.",
  "cf\\\\.", "uncultured", "unidentified", " clone", " enrichment"),
  verbose = TRUE, sleep = 0)
}
\arguments{
\item{taxon}{A taxon to get all species for}

\item{filters}{A character vector of strings to exclude}

\item{verbose}{Update on how many are done}

\item{sleep}{How many seconds to sleep between calls (on top of rentrez's defaults)}
}
\value{
A list containing a vector of dark names, a vector of known names, and fraction.dark
}
\description{
Separate dark from known taxa on another database
}
