% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspeciesinfo.R
\name{species_get_info}
\alias{species_get_info}
\title{Get information of a list of species}
\usage{
species_get_info(species)
}
\arguments{
\item{species}{A vector of names}
}
\value{
A data frame of info about species
}
\description{
Get information of a list of species
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or \url{https://eol.org/api/}
}
