% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtrees.R
\name{taxa_get_phylomatic_tree}
\alias{taxa_get_phylomatic_tree}
\title{Get phylomatic subtree}
\usage{
taxa_get_phylomatic_tree(taxa)
}
\arguments{
\item{taxa}{The vector of names, already resolved}
}
\value{
A phylo object
}
\description{
Get phylomatic subtree
}
\examples{
phy <- taxa_get_phylomatic_tree(c("Panthera leo", "Panthera onca",
         "Panthera tigris", "Panthera uncia"))
plot(phy)
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the interface of phylomatic \url{http://phylodiversity.net/phylomatic/}
}
