% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getspecies.R
\name{taxon_get_species}
\alias{taxon_get_species}
\title{Get all species from a taxon}
\usage{
taxon_get_species(taxon, filters = c("environmental", "sp\\\\.", "cf\\\\."))
}
\arguments{
\item{taxon}{A taxon to get all species for}

\item{filters}{A character vector of strings to exclude}
}
\value{
A vector of names
}
\description{
Get all species from a taxon
}
\examples{
print(taxon_get_species("Vulpes"))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription} or the rotl package, another interface to Open Tree of Life
}
