% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{text_get_scientific_names}
\alias{text_get_scientific_names}
\title{Function to pull scientific names from text}
\usage{
text_get_scientific_names(text)
}
\arguments{
\item{text}{The text string to extract names from}
}
\value{
A data.frame of scientific names and other data from GNRD
}
\description{
This takes a string of text and extracts any scientific names in the text. Other words in the text are ignored.
}
\examples{
text <- "Formica polyctena is a species of European red wood ant in
   the genus Formica. The pavement ant, Tetramorium caespitum
   is an ant native to Europe."
print(text_get_scientific_names(text))
}
