% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{text_get_scientific_names}
\alias{text_get_scientific_names}
\title{Function to pull scientific names from text}
\usage{
text_get_scientific_names(text, search_engine = 0)
}
\arguments{
\item{text}{The text string to extract names from}

\item{search_engine}{1 to use TaxonFinder, 2 to use NetiNeti, 0 to use both}
}
\value{
A vector of scientific names
}
\description{
Function to pull scientific names from text
}
\examples{
text <- "Formica polyctena is a species of European red wood ant in
   the genus Formica. The pavement ant, Tetramorium caespitum
   is an ant native to Europe."
print(text_get_scientific_names(text))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription}
}
