% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findscientificnames.R
\name{url_get_scientific_names}
\alias{url_get_scientific_names}
\title{Function to pull scientific names from web pages}
\usage{
url_get_scientific_names(URL, search_engine = 0)
}
\arguments{
\item{URL}{The URL to extract names from}

\item{search_engine}{1 to use TaxonFinder, 2 to use NetiNeti, 0 to use both}
}
\value{
A vector of scientific names
}
\description{
Function to pull scientific names from web pages
}
\examples{
URL <- "https://en.wikipedia.org/wiki/Plain_pigeon"
print(url_get_scientific_names(URL))
}
\seealso{
\url{https://github.com/phylotastic/phylo_services_docs/tree/master/ServiceDescription}
}
